Colored binary STL support
When colored faces are discovered in the STL file, they are groupped
into own Volumes to be printed with different extruders. Colors should
start from 0 (meaning first extruder) and each number represents
different extruder to use when printing the face. Volumes are filled
from 0 to maximal number of color used in the whole STL, so if the file
consists of only faces with color 1, second nozzle will be used to print
the model.
This functionality is not enabled by default and requires the -e flag to
be passed to CuraEngine